requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
Pillow
djangorestframework-simplejwt
dj-config-url
